refactor `curve` adaptor
pull data dynamically from api.curve.fi
- pool info
- pool apy from subgraph proxy on api.curve.fi

keep legacy code needed for `curve-finance` adaptor
